split out join hash map from hash map
